requirements.txt
astroid==2.4.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-dotenv==0.15.0
six==1.15.0
slack-bolt==1.0.0rc2
slack-sdk==3.0.0rc3
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
